[src styles] color white for id=app
i.e. all elements inside the main app element (where 'everything' is) will
be white by default, as I'm in the future adding more elements to the
app/page, beyond the OP cards themselves.  a subsequet commit is needed
to change the 'color' of OP cards as they are now inherting this white color
but will need black color as their elements bg is white.