[src jsx] separate 'display' and 'sagas' modes picker; improve modePicker
where 'sagas' gets sourced from the config and uses the shared 'sagas'
JS object stash that contains all the sagas.

also some more features (onClick, etc) around the components, there's
still much more work to do and it's clear I'll need to bring in some
state management library (e.g. redux) as to have the modePicker
component work neatly and also to 'communicate' to the OP cards
about state changes, so that "filters" feature can be introduced to the
app.

idea is that the modePickers for 'Sagas' dictate which sagas the User
can see facts for.