[src app] setup filtering logic for facts
facts listed now are dependent on the picked 'sagas' on the top.  this
is done via react-redux.  A <Provider> component is added that wraps the
whole of the app with a 'store' prop so that state can be shared through
all React components and that reducers can be used based on dispatched
actions.

filters are changed through reducers and their state maintained by
Redux.